+++ lib/lean_types.dart
@@ -38,7 +38,11 @@ enum Params {
   show_consent_explanation,
   destination_alias,
   destination_avatar,
-  consent_id
+  consent_id,
+  consent_attempt_id,
+  granular_status_code,
+  status_additional_info,
+  customer_metadata
 }
 
 enum LeanMethods {
@@ -51,6 +55,9 @@ enum LeanMethods {
   updatePaymentSource,
   verifyAddress,
   authorizeConsent,
+  checkout,
+  manageConsents,
+  captureRedirect,
 }
 
 enum LeanPermissions {
@@ -102,6 +109,7 @@ class LeanResponse {
   final String? exitIntentPoint;
   final String? exitSurveyReason;
   final String? leanCorrelationId;
+  final String? userExitIntent;
 
   LeanResponse({
     required this.status,
@@ -114,5 +122,6 @@ class LeanResponse {
     this.exitIntentPoint,
     this.exitSurveyReason,
     this.leanCorrelationId,
+    this.userExitIntent,
   });
 }