+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 3.0.12
+- Added new sdk methods - `checkout`, `manageConsents`, `captureRedirect`
+- Added support for `user_exit_intent` in callback response
+
 ## 3.0.11
 
 Added new sdk method - `authorizeConsent`